+++ src/onegov.wtfs/onegov/wtfs/assets/js/iframe-resizer-options.js
@@ -1,8 +1,0 @@
-window.iFrameResizer = {
-    heightCalculationMethod: function() {
-        return $('header').outerHeight(true) + $('main').height();
-    },
-    readyCallback: function() {
-        parentIFrame.scrollTo(0, 0);
-    }
-};